+++ packages/terra-date-time-picker/src/DateTimeUtils.js
@@ -1,5 +1,6 @@
 import moment from 'moment-timezone';
 import DateUtil from 'terra-date-picker/lib/DateUtil';
+import TimeUtil from 'terra-time-input/lib/TimeUtil';
 
 class DateTimeUtils {
   /**
@@ -29,47 +30,45 @@ class DateTimeUtils {
   }
 
   /**
-   * Converts an ISO string to the given format.
-   * @param {string} iSODate - The ISO string to convert.
-   * @param {string} format - The desired date/time format for the conversion
-   * @return {string} - The formatted date/time string.
-   */
-  static formatISODateTime(iSODate, format) {
-    if (!iSODate || iSODate.length <= 0) {
-      return '';
-    }
-
-    const momentDate = moment(iSODate);
-    return DateTimeUtils.formatMomentDateTime(momentDate, format);
-  }
-
-  /**
-   * Converts a moment object to the given format.
-   * @param {object} momentDate - The moment object to convert.
-   * @param {string} format - The desired date/time format for the conversion
-   * @return {string} - The formatted date/time string.
-   */
-  static formatMomentDateTime(momentDate, format) {
-    return momentDate && momentDate.isValid() ? momentDate.format(format) : undefined;
-  }
-
-  /**
-   * Synchronize the date and time for a given moment object.
+   * Synchronize the date and time for a given the base moment object.
+   * Because a moment object must have the date portion, if the provided iSOdate is invalid,
+   * The base moment object will not synchronize the time if the provided time is invalid.
    * @param {object} momentDate - The moment object to synchronize the date and time.
-   * @param {string} iOSdate - The date string to synchronize with the moment object.
+   * @param {string} iSOdate - The date string to synchronize with the moment object.
    * @param {string} time - The time to synchronize with the moment object.
    * @return {object} - The synchronized moment object.
    */
-  static syncDateTime(momentDate, iOSdate, time) {
-    const date = moment(iOSdate);
-    let newDate = momentDate ? momentDate.clone() : date;
+  static syncDateTime(momentDate, iSOdate, time) {
+    const date = moment(iSOdate);
 
-    // If momentDate was null, a new moment date needs to be created and sync'd with the entered time.
-    if (momentDate === null && time && time.length === 5) {
-      newDate = DateTimeUtils.updateTime(newDate, time);
+    // If the base momentDate is valid, sync the date and time is they are valid.
+    if (momentDate && momentDate.isValid()) {
+      let tempDate = momentDate.clone();
+
+      if (date.isValid()) {
+        tempDate.year(date.get('year')).month(date.get('month')).date(date.get('date'));
+      }
+
+      if (time && time.length === 5) {
+        tempDate = DateTimeUtils.updateTime(tempDate, time);
+      }
+
+      return tempDate;
     }
 
-    return newDate.year(date.get('year')).month(date.get('month')).date(date.get('date'));
+    // If the base momentDate is invalid, use the iSOdate as the base and update the time if valid.
+    if (date.isValid()) {
+      let tempDate = date.clone();
+
+      if (time && time.length === 5) {
+        tempDate = DateTimeUtils.updateTime(tempDate, time);
+      }
+
+      return tempDate;
+    }
+
+    // Neither the base momentDate nor the iSOdate is valid.
+    return momentDate;
   }
 
   /**
@@ -97,18 +96,7 @@ class DateTimeUtils {
    * @return {boolean} - True if both the date and time are valid and conform to the format.
    */
   static isValidDateTime(date, time, format) {
-    return DateTimeUtils.isValidDate(date, format) && DateTimeUtils.isValidTime(time);
-  }
-
-  /**
-   * Determines if the date is valid and conforms to the given format.
-   * @param {string} date - The date to validate.
-   * @param {string} format - The date format to use for the validation.
-   * @return {boolean} - True if the date is valid and conforms to the format.
-   */
-  static isValidDate(date, format) {
-    const dateMoment = moment(date, format, true);
-    return dateMoment.isValid();
+    return DateUtil.isValidDate(date, format) && DateTimeUtils.isValidTime(time);
   }
 
   /**
@@ -173,5 +161,8 @@ class DateTimeUtils {
     return DateTimeUtils.updateTime(DateUtil.createSafeDate(DateUtil.convertToISO8601(date, dateformat)), time);
   }
 }
+
+DateTimeUtils.FORMAT_12_HOUR = TimeUtil.FORMAT_12_HOUR;
+DateTimeUtils.FORMAT_24_HOUR = TimeUtil.FORMAT_24_HOUR;
 
 export default DateTimeUtils;