+++ packages/terra-dialog-modal/tests/jest/DialogModalExample.jsx
@@ -9,11 +9,6 @@ import DialogModal from '../../src/DialogModal';
 
 const propTypes = {
   width: PropTypes.string,
-  isFocused: PropTypes.bool,
-};
-
-const defaultProps = {
-  isFocused: false,
 };
 
 class DialogModalExample extends React.Component {
@@ -51,7 +46,6 @@ class DialogModalExample extends React.Component {
         <DialogModal
           ariaLabel="Dialog Modal Example"
           isOpen={this.state.isOpen}
-          isFocused={this.props.isFocused}
           onRequestClose={this.handleCloseModal}
           header={<ActionHeader title="Action Header used here" onClose={this.handleCloseModal} />}
           footer={<ActionFooter start="Footer Goes here" />}
@@ -66,5 +60,4 @@ class DialogModalExample extends React.Component {
 }
 
 DialogModalExample.propTypes = propTypes;
-DialogModalExample.defaultProps = defaultProps;
 export default DialogModalExample;