+++ packages/terra-modal-manager/src/terra-dev-site/test/modal-manager/ModalManagerIntegration.test.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import DisclosureComponent from './DisclosureIntegrationComponent';
+import ModalManager from '../../../ModalManager';
+
+const ModalManagerDefault = () => (
+  <div role="main" style={{ height: '100%' }}>
+    <ModalManager>
+      <DisclosureComponent identifier="root-component" disclosureType="modal" />
+    </ModalManager>
+  </div>
+);
+
+export default ModalManagerDefault;