+++ packages/terra-layout/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+### Breaking Changes
+* Update react-router-dom peerDependency to latest stable major release
+* Updated tabbable dependency to ^4.0.0
 
 3.4.0 - (April 24, 2019)
 ------------------