+++ packages/terra-time-input/CHANGELOG.md
@@ -3,10 +3,19 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Change the onBlur event to get triggered only when the entire time input component loses focus.
+
+### Added
+* Added 'onFocus' prop.
+
+### Removed
+* Removed the 'onInputFocus' custom prop in favor of the first-class 'onFocus' prop.
 
 3.7.0 - (April 24, 2019)
 ------------------
 ### Changed
+* Replaced string "12-hour" with TimeUtil.FORMAT_12_HOUR in terra dev-site eamples
 * Moved packages we use solely in doc examples and tests to devDependencies
 
 3.6.0 - (March 26, 2019)